gitlab: Upgrade gitlab-shell & gitlab-workhorse to versions which propagate $HOME
As was described in the previous patch, we need $HOME to be propagated
by this programs so that git can find partition's .gitconfig.

Specifically we need the following patches to be present in our build:

    https://gitlab.com/gitlab-org/gitlab-shell/commit/9e087f64
    https://gitlab.com/gitlab-org/gitlab-workhorse/commit/b5f1b803

They both have been applied upstream very close to revisions we
previously had in software.cfg, so we only need to update the revisions
to get them.

/cc @kazuhiko, @jerome