apache frontend: describe prefer-gzip-encoding-to-backend
From Cédric message

When a client do a request, you pass the Accept-Encoding header. For me it is:
```
Accept-Encoding:gzip, deflate, sdch, br
```

But in a lot of case the cache configuration has `Vary:Accept-Encoding` . `Accept-Encoding` request header may vary a lot from a client to another but one of the common denominator is `gzip`. So in order to avoid a multiplication of cache versions in CDN, if `gzip` is present in the `Accept-Encoding` request header all other format are stripped from the request in order to reduce variation in Trafficserver.

When this parameter is activated:
```
Accept-Encoding:gzip, deflate, sdch, br
```
is transmitted as
```
Accept-Encoding:gzip
```
And
```
Accept-Encoding:deflate, sdch, br
```
is transmitted as
```
Accept-Encoding:deflate, sdch, br
```