jstestnode: parsing results with lxml, ios version and appium url parameters
With the introduction of iOS test support tests we're taking very long to run
in iOS. This happens because of the required architecture and Selenium Webdriver.
To avoid this issue now we parse the results from the webpage using lxml. This
way we can close the browser as soon as tests finish to run.

Also command line options to a specific iOS version and to change the Appium
url were implemented. Sauce Labs paid service for remote iOS testing is supported.