jstestnode: add support to send appium parameters as slapos parameters
This allows user to customize in the test suite module on which system
they want to run the tests (Firefox or iOS) through the slapos parameters.
In iOS case, it's possible to change the iOS version and it's required
that the user give the SauceLabs credentials in form of user:apikey using
the `appium_server_auth` parameter.