pycurl: adds the component when needed to fulfill the new dependency of slapos.toolbox
Slapos.toolbox has a new dependency to pycurl, which fails to compile if libcurl cannot be found. This commit adds the component/pycurl dependency to most of the Software Releases which have a dependency to slapos.toolbox.

I tried to do it globally, so I couldn't check the compilation of all the modified SRs. Is it a way to test that all SRs still compiled ?

Also, did I forget some important SR ?

@rafael @alain.takoudjou @kazuhiko @jm 

/reviewed-on https://lab.nexedi.com/nexedi/slapos/merge_requests/97