erp5: jupyter.enable is boolean, not string
When originally in 0a446263 (ERP5 and Jupyter integrated together) I added
Jupyter support into ERP5 the parameter for whether to enable/disable it was
declared as boolean in JSON schema but processed as string in instance code,
this way preventing usage of real JSON's boolean.

Fix it.

(also fixing up software/slapos-master/ which copied the files for play in 87d13789)

/noticed-by @vpelletier (on https://lab.nexedi.com/nexedi/slapos/merge_requests/43)