jstestnode: added support for running tests with Selenium Remote and Appium
It uses Appium, which provides a Selenium WebDrivercompatible API to remotely control an iOS (or Android) simulator.This way we can run tests in both mobile OSes without big changesto the current test code and infrastructure.

This allows user to customize in the test suite module on which systemthey want to run the tests (Firefox or iOS) through the slapos parameters.In iOS, for example, it's possible to change the iOS version and it's requiredthat the user give the SauceLabs credentials in form of user:apikey usingthe `appium_server_auth` parameter.

An example of parameters to use in a this suite:

```
{"mariadb": {  "relaxed-writes": true,  "mariadb-relaxed-writes": true,  "test-database-amount": 30},  "target": "ios",  "target-version": "9.3",  "appium-server-auth": "username:api_key  "test-suite" : "jio" ,  "test-url": "jio-repository.git/test/tests.html" }
```