component/ocropy Add the ocropy library for OCR
This commit add component needed for Optical Character recognition:
ocropy and its dependancies

The Ocropy egg on Pypi is empty, so this component have to be built from
source. The default neural network model is downloaded from the author
website and installed in python2.7/share/ocropus.

The patch is here for to different thing: one is adding an
OCROPY_MODEL_PATH to be able to specify the position of models, the
other is importing for the first time a module that compile C code from
python, allowing us to have the static library built before runtime and
preventing us from using gcc through zope.

More models are available here:
https://github.com/tmbdev/ocropy/wiki/Models