component/ocropy Add the ocropy library for OCR
This commit add component needed for Optical Character recognition:
ocropy and its dependancies

The Ocropy egg on Pypi is empty, so this component have to be built from
source. The default neural network model is downloaded from the author
website and installed in python2.7/share/ocropus.

More models are available here:
https://github.com/tmbdev/ocropy/wiki/Models