apache-frontend: Extend nginx slave configuration to be possible restrict the access by password and/or address
By default authentification by password is enabled. To disable just pass --enable-auth=false

By default all addresses are allowed. if 'allowed-host-list' is passed we set "allow" to the addresses passed as parameter and deny all the rest.

Note: site_url is not published anymore because it has the same information of url