component add tesseract 4.0 and dependancys to slapos
This commit contain tesseract4.0 and its up do date leptonica
dependancy as well as a python wrapper that use tesseract API to do text
recognition.

Currently, the wrapper cannot initialise the api as intended and the
traineddata emplacement is not set (an new environnement variable should
be introduced in zope, but it should be possible to "force" a new path
with a patch that change the tesserocr _DEFAULT_PATH value.