slapos_crm: Include Report for rank Computers by instance error percentage
This report lists the Computers that host at least one instance of a cluster which is returning error (see Related ticket columns). This list also the utilisation ratio of the partitions, and as well the percentage of the partitions inside which are reporting errors.

This report aims to help on Support Decision Process to eliminate the biggest amount of errors.