Fixed Jupyter's Unicode file issues by setting the process' locale
@luke @Tyagov 

When SlapOS launches services they don't have their locale set. This was causing problems with unicode characters in Jupyter notebook names. For example, calling the notebook "Ł" crashes Jupyter's kernel. Hardcoding the UTF-8 locale to the ipython_notebook script fixes that but we should look into making a generic fix for any other services that might have this issue in the future.

/reviewed-on https://lab.nexedi.com/nexedi/slapos/merge_requests/193