caucase: allow to defined how many csr should be signed automatically
The caucase parameter 'auto-sign-csr' used to say if the first csr should be signed
was changed to 'auto-sign-csr-amount' to let specify how many csr can be signed without
human intervention.
This parameter will be set to 2 in erp5 software release, so that caucase http server and
apache server cas request thier certificate automatically.