component/firefox: simplify profile
Major change is that slapos compatible wrapper is no longer installed in
parts/firefox/firefox-slapos , but directly as "firefox" in the
buildout:bin-directory of the software profile.
geckodriver is also in the same buildout:bin-directory.
This way, softwares using this just need to add
buildout:bin-directory to $PATH

erp5testnode knows since
https://lab.nexedi.com/nexedi/erp5/commit/59ee7970bdc0ea6efe7b16b8953042f675f82e78
that it should look for firefox in buildout:bin-directory of the
installed SR, so we should not need backward compatibility here.

Adjust seleniumrunner and jstestnode for the changes

Also drop unused firefox 45

Jérome changes from original 8cf6908df752b9e7632b2910ffc85027f1f4f22c :
 - name sections [firefox-wrapper] and [firefox] because including the
version in section name means we have to also update section name when
we update version. Users have to be careful of installing
${firefox-wrapper:} and not ${firefox:}
 - introduce macros for implementation and simple section using the
macros, to make it easy to had more versions or to increase versions,
for both firefox and geckodriver. They are slightly different because
for firefox we need a wrapper, but geckodriver is usable directly.
 - use same version, the "versions up" will be separate commits.
 - now that seleniumrunner and jstestnode are using buildout.hash.cfg,
md5sums are in separate files.

Co-authored-by: Jérome Perrin <jerome@nexedi.com>