neotest: Actually add instance and teach it to run `neotest bench-local` under Nexedi testing infrastructure
Commit 31fe231f (neotest: Draft software-release to run NEO/go & friends
tests/benchmarks under webrunner) added "neotest" software release, but
there it added only software supplying - not instantiation - this way
allowing only manual play with neotest.

Recently @rafael approached me saying that one of my webrunner is red on
monitoring because "neotest" software inside fails to instantiate, and
suggested to create a simple instance with bin/runTestSuite inside, so
that this way neotest could be added to a testnode and run on regular
basis by our testing infrastructure.

So here it goes - simple instance with only bin/runTestSuite which tries
to adhere to testnode protocol.

For the actual runTestSuite program I followed the same approach as for
bin/neotest itself - the main program is not processed via any templates
and is used as it was pristinely downloaded, and all buildout
integration is done via first sourcing $SR/neotest-env.sh .
So bin/runTestSuite inside instance follows this approach.

erp5.util is updated from 0.4.49 to latest 0.4.50 because there were
backward-incompatible changes in 0.4.50 release (2 @luke: which btw will
render neoppod and build-rina runTestSuite(s) broken when they try to
upgrade) and for me it would be a waste of time to later go through
updating for 0.4.49 -> 0.4.50 API change, so I used the latest release
right from start.

/cc @Tyagov
/see also https://lab.nexedi.com/kirr/neo/commit/526491d3