grafana: make influxdb listen on local ip for RPC protocol
See
https://docs.influxdata.com/influxdb/v1.4/administration/config/#bind-address-127-0-0-1-8088

This should not be reachable from outside because it can be used to pull
a backup.
Because we usually have reachable ipv6 and local only ipv4, make it
ipv4 and add a note in README because this is not always true.

Also make ports use explicit names in profile and fix promise that
should check http-port and not rpc-port