golang: Hook-in pkg-config for Cgo support
The most convenient way to discover CFLAGS/LDFLAGS when building Cgo bits is
via pkg-config:

    https://golang.org/cmd/cgo/

For this to work let's add pkg-config to be available there in gowork out of
the box, and provide users with a way to specify which C packages they want to
be there on gowork's $PKG_CONFIG_PATH, e.g. this way:

    [gowork]
    cpkgpath =
        ${sqlite3:location}/lib/pkgconfig
        ${zlib:location}/lib/pkgconfig

In buildout sources cpkgpath is specified as multiline - not one line with ":"
delimiters - with the idea that it should be easy to amend gowork.cpkgpath with "+="
in different places.